Address PR review feedback
- Renamed `set` to `managedSeedSet`.
- Move predicates to add.go as it is done for other controllers.
- Add `generationChange` predicate to `ManagedSeedSet` object.
