+++ pkg/controllermanager/controller/managedseedset/reconciler.go
@@ -43,8 +43,8 @@ type Reconciler struct {
 func (r *Reconciler) Reconcile(ctx context.Context, request reconcile.Request) (reconcile.Result, error) {
 	log := logf.FromContext(ctx)
 
-	set := &seedmanagementv1alpha1.ManagedSeedSet{}
-	if err := r.Client.Get(ctx, request.NamespacedName, set); err != nil {
+	managedSeedSet := &seedmanagementv1alpha1.ManagedSeedSet{}
+	if err := r.Client.Get(ctx, request.NamespacedName, managedSeedSet); err != nil {
 		if apierrors.IsNotFound(err) {
 			log.V(1).Info("Object is gone, stop reconciling")
 			return reconcile.Result{}, nil
@@ -52,34 +52,33 @@ func (r *Reconciler) Reconcile(ctx context.Context, request reconcile.Request) (
 		return reconcile.Result{}, fmt.Errorf("error retrieving object from store: %w", err)
 	}
 
-	if set.DeletionTimestamp != nil {
-		return r.delete(ctx, log, set)
+	if managedSeedSet.DeletionTimestamp != nil {
+		return r.delete(ctx, log, managedSeedSet)
 	}
-	return r.reconcile(ctx, log, set)
+	return r.reconcile(ctx, log, managedSeedSet)
 }
 
-func (r *Reconciler) reconcile(ctx context.Context, log logr.Logger, set *seedmanagementv1alpha1.ManagedSeedSet) (result reconcile.Result, err error) {
+func (r *Reconciler) reconcile(ctx context.Context, log logr.Logger, managedSeedSet *seedmanagementv1alpha1.ManagedSeedSet) (result reconcile.Result, err error) {
 	// Ensure gardener finalizer
-	if !controllerutil.ContainsFinalizer(set, gardencorev1beta1.GardenerName) {
+	if !controllerutil.ContainsFinalizer(managedSeedSet, gardencorev1beta1.GardenerName) {
 		log.Info("Adding finalizer")
-		if err := controllerutils.AddFinalizers(ctx, r.Client, set, gardencorev1beta1.GardenerName); err != nil {
+		if err := controllerutils.AddFinalizers(ctx, r.Client, managedSeedSet, gardencorev1beta1.GardenerName); err != nil {
 			return reconcile.Result{}, fmt.Errorf("could not add finalizer: %w", err)
 		}
-		return reconcile.Result{}, nil
 	}
 
 	var status *seedmanagementv1alpha1.ManagedSeedSetStatus
 	defer func() {
 		// Update status, on failure return the update error unless there is another error
-		if updateErr := r.updateStatus(ctx, set, status); updateErr != nil && err == nil {
+		if updateErr := r.updateStatus(ctx, managedSeedSet, status); updateErr != nil && err == nil {
 			err = fmt.Errorf("could not update status: %w", updateErr)
 		}
 	}()
 
 	// Reconcile creation or update
 	log.V(1).Info("Reconciling creation or update")
-	if status, _, err = r.Actuator.Reconcile(ctx, log, set); err != nil {
-		return reconcile.Result{}, fmt.Errorf("could not reconcile ManagedSeedSet %s creation or update: %w", kutil.ObjectName(set), err)
+	if status, _, err = r.Actuator.Reconcile(ctx, log, managedSeedSet); err != nil {
+		return reconcile.Result{}, fmt.Errorf("could not reconcile ManagedSeedSet %s creation or update: %w", kutil.ObjectName(managedSeedSet), err)
 	}
 	log.V(1).Info("Creation or update reconciled")
 
@@ -87,9 +86,9 @@ func (r *Reconciler) reconcile(ctx context.Context, log logr.Logger, set *seedma
 	return reconcile.Result{RequeueAfter: r.Config.SyncPeriod.Duration}, nil
 }
 
-func (r *Reconciler) delete(ctx context.Context, log logr.Logger, set *seedmanagementv1alpha1.ManagedSeedSet) (result reconcile.Result, err error) {
+func (r *Reconciler) delete(ctx context.Context, log logr.Logger, managedSeedSet *seedmanagementv1alpha1.ManagedSeedSet) (result reconcile.Result, err error) {
 	// Check gardener finalizer
-	if !controllerutil.ContainsFinalizer(set, gardencorev1beta1.GardenerName) {
+	if !controllerutil.ContainsFinalizer(managedSeedSet, gardencorev1beta1.GardenerName) {
 		log.V(1).Info("Skipping as it does not have a finalizer")
 		return reconcile.Result{}, nil
 	}
@@ -100,7 +99,7 @@ func (r *Reconciler) delete(ctx context.Context, log logr.Logger, set *seedmanag
 		// Only update status if the finalizer is not removed to prevent errors if the object is already gone
 		if !removeFinalizer {
 			// Update status, on failure return the update error unless there is another error
-			if updateErr := r.updateStatus(ctx, set, status); updateErr != nil && err == nil {
+			if updateErr := r.updateStatus(ctx, managedSeedSet, status); updateErr != nil && err == nil {
 				err = fmt.Errorf("could not update status: %w", updateErr)
 			}
 		}
@@ -108,16 +107,16 @@ func (r *Reconciler) delete(ctx context.Context, log logr.Logger, set *seedmanag
 
 	// Reconcile deletion
 	log.V(1).Info("Reconciling deletion")
-	if status, removeFinalizer, err = r.Actuator.Reconcile(ctx, log, set); err != nil {
-		return reconcile.Result{}, fmt.Errorf("could not reconcile ManagedSeedSet %s deletion: %w", kutil.ObjectName(set), err)
+	if status, removeFinalizer, err = r.Actuator.Reconcile(ctx, log, managedSeedSet); err != nil {
+		return reconcile.Result{}, fmt.Errorf("could not reconcile ManagedSeedSet %s deletion: %w", kutil.ObjectName(managedSeedSet), err)
 	}
 	log.V(1).Info("Deletion reconciled")
 
 	// Remove gardener finalizer if requested by the actuator
 	if removeFinalizer {
-		if controllerutil.ContainsFinalizer(set, gardencorev1beta1.GardenerName) {
+		if controllerutil.ContainsFinalizer(managedSeedSet, gardencorev1beta1.GardenerName) {
 			log.Info("Removing finalizer")
-			if err := controllerutils.RemoveFinalizers(ctx, r.Client, set, gardencorev1beta1.GardenerName); err != nil {
+			if err := controllerutils.RemoveFinalizers(ctx, r.Client, managedSeedSet, gardencorev1beta1.GardenerName); err != nil {
 				return reconcile.Result{}, fmt.Errorf("failed to remove finalizer: %w", err)
 			}
 		}
@@ -129,11 +128,11 @@ func (r *Reconciler) delete(ctx context.Context, log logr.Logger, set *seedmanag
 	return reconcile.Result{RequeueAfter: r.Config.SyncPeriod.Duration}, nil
 }
 
-func (r *Reconciler) updateStatus(ctx context.Context, set *seedmanagementv1alpha1.ManagedSeedSet, status *seedmanagementv1alpha1.ManagedSeedSetStatus) error {
+func (r *Reconciler) updateStatus(ctx context.Context, managedSeedSet *seedmanagementv1alpha1.ManagedSeedSet, status *seedmanagementv1alpha1.ManagedSeedSetStatus) error {
 	if status == nil {
 		return nil
 	}
-	patch := client.StrategicMergeFrom(set.DeepCopy())
-	set.Status = *status
-	return r.Client.Status().Patch(ctx, set, patch)
+	patch := client.StrategicMergeFrom(managedSeedSet.DeepCopy())
+	managedSeedSet.Status = *status
+	return r.Client.Status().Patch(ctx, managedSeedSet, patch)
 }
